Add support for declaring unmeasured data
Add recognition of a special ELF section, '.sgx.unmeasured', that can
be used by enclaves to declare data that should not contribute to the
measurement of an enclave.  The section's address and size must both
be page aligned.  Pages that are wholly contained by the unmeasured
sections are built with ADD_PAGE_ONLY, i.e. they skip EEXTEND.

Report an error if the unmeasured section is not properly aligned, or
if the section contains any initialized data.  Assert if the section
straddles multiple segments (invalid ELF file).

Define the macros '__unmeasured' and '__unmeasured_aligned' for use
by enclave developers in declaring unmeasured data.  The aligned macro
is intended to be used once per enclave to ensure proper alignment.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>