Add --no-daemon option to AESM
Application container runtimes, e.g. Docker and RKT, do not provide a
program manager, e.g. systemd, inside the container, and so attempting
to run AESM as a daemon will fail.   Add --no-daemon to allow running
the AESM in a container as a normal process, i.e. do not call daemon().

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>