Fix potential false alarm of stack smashing attack
This commit makes sure all TRTS functions during which the stack guard
(or canary value) may be initialized are compiled without
-fstack-protector-strong. Currently, functions (in trts_nsp.cpp) that
directly deal with stack guard initialization are correctly compiled
without -fstack-protector-strong; yet, the caller of these functions
(do_ecall() in trts_ecall.cpp and enter_enclave() in trts.cpp) are not.
This makes the code of enter_enclave() and do_ecall() very fragile since
any code changes that cause the compiler to add stack protection logic
for these two functions would break them. This commit simply moves the
code of do_ecall() and enter_enclave() to trts_nsp.cpp. As the name of
the file trts_nsp.cpp suggests, it is where No-Stack-Protection code are
supposed to be anyway.

Signed-off-by: Hongliang Tian <hongliang.tian@intel.com>