WIP: figure more general mechanism to build ipp crypto
When building SDK & IPP Crypto from source with nix, the current check
for the IPP source code does not work properly. That is, even though the
IPP Crypto source code is there the check in the Makefile triggers the
call to the ipp_source target which invokes git to update the submodule
or clone the repo.

In the context of a nix derivation, the source code along with the
submodules is fetch at the beginning and the .git directory is removed,
such that later invocations to git commands such as the one in the
Makefile for the ipp_source target, fail.

Also, depending on the environment, NIX_PATH may not be set and is
therefore too restrictive as a check or assumes too much. A more general
mechanism would be preferable.

Signed-off-by: Sylvain Bellemare <sbellem@gmail.com>