Disable use of undefined DEF_WEAK macro in strptime.c
The strptime.c code invokes, what is presumably an undefined macro
called, DEF_WEAK. The compiler can't resolve this to a function, so
being traditional C, it invents a function taking arbitrary args and
returning an int.

GCC 9.5.0 used in the reproducible build warns about this problem
but sdk/Makefile.source is throwing stderr from the tlibc build
by sending it all to /dev/null, so the warning is never seen.

  time/strptime.c:92:1: warning: data definition has no type or storage class
     92 | DEF_WEAK(strptime);
        | ^~~~~~~~
  time/strptime.c:92:1: warning: type defaults to 'int' in declaration of 'DEF_WEAK' [-Wimplicit-int]
  time/strptime.c:92:1: warning: parameter names (without types) in function declaration

New GCC 14 has turned this warning into a fatal error, such that
this bug breaks the build outside of the NixOS envirnoment.

The other use of DEF_WEAK in tlibc/time/localtime.c was surrounded
by '#if 0', so do the same in strptime.c

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>